requirements.txt
torch>=1.13.0
# pytorch-geometric>=2.3.0
transformers>=4.30.0
diffusers>=0.16.0
accelerate>=0.20.0
numpy>=1.24.0
trimesh>=3.22.0
matplotlib>=3.7.0
pillow>=9.5.0
tqdm>=4.65.0
open3d>=0.17.0
scikit-learn>=1.2.0
scipy>=1.10.0
django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
channels>=4.0.0
daphne>=4.0.0
whitenoise>=6.4.0
# pytorch3d>=0.7.2
# kaolin>=0.14.0
plotly>=5.14.0
wandb>=0.15.0
tensorboard>=2.13.0
